Сгенерированные Q-классы и исправленные дао после задач https://jira.aplana.com/browse/SBRFNDFL-2075 и https://jira.aplana.com/browse/SBRFNDFL-2078

+++ dao/src/main/java/com/aplana/sbrf/taxaccounting/dao/impl/refbook/RefBookAsnuDaoImpl.java
@@ -45,7 +45,7 @@ public class RefBookAsnuDaoImpl implements RefBookAsnuDao {
      * @return Список значений справочника
      */
     @Override
-    public List<RefBookAsnu> fetchByIds(List<Integer> ids) {
+    public List<RefBookAsnu> fetchByIds(List<Long> ids) {
         return sqlQueryFactory
                 .select(refBookAsnuBean)
                 .from(refBookAsnu)
